refactor(web): move legacy API to web.v0 package
Introduce a web package hierarchy that makes the structural intent clear
and leaves room for a future web.v1 package alongside it.

- rest.api.* -> web.v0.api.*
- rest.controller.* -> web.v0.controller.*
- ChannelFinderExceptionHandler -> web (cross-cutting, not legacy-specific)

Update all test imports accordingly. No behaviour change.
