+++ src/test/java/org/phoebus/channelfinder/docker/ChannelFinderChannelsIT.java
@@ -29,7 +29,7 @@ import org.junit.jupiter.api.Test;
 import org.phoebus.channelfinder.common.CFResourceDescriptors;
 import org.phoebus.channelfinder.docker.ITUtil.AuthorizationChoice;
 import org.phoebus.channelfinder.entity.Channel;
-import org.phoebus.channelfinder.rest.controller.ChannelController;
+import org.phoebus.channelfinder.web.v0.controller.ChannelController;
 import org.testcontainers.containers.ComposeContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
