+++ src/test/java/org/phoebus/channelfinder/epics/EpicsRPCRequestIT.java
@@ -19,10 +19,10 @@ import org.phoebus.channelfinder.common.NTXmlUtil;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.Tag;
-import org.phoebus.channelfinder.rest.api.IChannel;
-import org.phoebus.channelfinder.rest.api.IProperty;
-import org.phoebus.channelfinder.rest.api.ITag;
 import org.phoebus.channelfinder.service.ChannelFinderEpicsService;
+import org.phoebus.channelfinder.web.v0.api.IChannel;
+import org.phoebus.channelfinder.web.v0.api.IProperty;
+import org.phoebus.channelfinder.web.v0.api.ITag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.security.test.context.support.WithMockUser;
